requirements.txt
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.58.5
fsspec==2024.6.1
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
lpips==0.1.4
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==25.0
pillow==11.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scikit-image==0.24.0
scipy==1.13.1
six==1.17.0
sympy==1.13.3
tifffile==2024.8.30
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
typing_extensions==4.12.2
zipp==3.23.0
